+++ src/components/TasksTable/Chronological.tsx
@@ -1,16 +1,19 @@
-import Table from "@mui/material/Table";
-import TableBody from "@mui/material/TableBody";
-import TableCell from "@mui/material/TableCell";
-import TableContainer from "@mui/material/TableContainer";
-import TableHead from "@mui/material/TableHead";
-import TableRow from "@mui/material/TableRow";
-import Paper from "@mui/material/Paper";
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableRow,
+  Paper,
+  Box,
+} from "@mui/material";
+import { useConfirm } from "material-ui-confirm";
+
 import { Task } from "../../hooks/useTasks";
 import { formatDuration, dateToStrTime } from "../../utils/dates";
-import { useConfirm } from "material-ui-confirm";
 import SyncIndicator from "./SyncIndicator";
 import CopyToClipboardBtn from "../CopyToClipboardBtn";
-import Box from "@mui/material/Box";
 import EditDeleteActions from "./EditDeleteActions";
 import StartStopActions from "./StartStopActions";
 import FavoriteAction from "./FavouriteAction";
@@ -41,9 +44,7 @@ const Chronological = ({
 }) => {
   const confirm = useConfirm();
   const deleteHandler = (id: number) => {
-    confirm({ description: "Are you sure you want to delete this task?" }).then(
-      () => deleteTask(id),
-    );
+    confirm({ description: "Delete task?" }).then(() => deleteTask(id));
   };
 
   return (