+++ src/components/TasksTable/Grouped.tsx
@@ -1,14 +1,18 @@
 import { useState, useEffect } from "react";
-import Table from "@mui/material/Table";
-import TableBody from "@mui/material/TableBody";
-import TableCell from "@mui/material/TableCell";
-import TableContainer from "@mui/material/TableContainer";
-import TableHead from "@mui/material/TableHead";
-import TableRow from "@mui/material/TableRow";
-import Paper from "@mui/material/Paper";
+
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableRow,
+  Paper,
+  Button,
+} from "@mui/material";
+
 import { Task } from "../../hooks/useTasks";
 import { formatDuration } from "../../utils/dates";
-import Button from "@mui/material/Button";
 import GroupedModal from "./GroupedModal";
 import SyncIndicator from "./SyncIndicator";
 import CopyToClipboardBtn from "../CopyToClipboardBtn";