+++ src/components/TasksTable/SyncIndicator.tsx
@@ -1,6 +1,7 @@
-import { Task } from "../../hooks/useTasks";
 import CloudDoneIcon from "@mui/icons-material/CloudDone";
 import CloudOffIcon from "@mui/icons-material/CloudOff";
+
+import { Task } from "../../hooks/useTasks";
 
 const SyncIndicator = ({ task }: { task: Task }) =>
   task.reported ? <CloudDoneIcon /> : <CloudOffIcon />;