+++ src/components/TasksTable/TasksTable.tsx
@@ -1,7 +1,9 @@
 import { useContext } from "react";
+
+import { Box } from "@mui/material";
+
 import Chronological from "./Chronological";
 import Grouped from "./Grouped";
-import Box from "@mui/material/Box";
 import { Task } from "../../hooks/useTasks";
 import { SettingsContext } from "../Settings/Provider";
 import { ViewType } from "../../hooks/useSettings";